requirements.txt
python-telegram-bot==20.7
google-generativeai==0.3.1
python-dotenv==1.0.0
pymongo==4.6.1
gunicorn==21.2.0